+++ backend/.env.example
@@ -1,0 +1,30 @@
+# TechBlogHub Backend Environment Variables Template
+# Copy this file to .env and fill in your values for local development
+
+# Database Configuration (for docker/production profiles)
+DATABASE_URL=jdbc:postgresql://localhost:5432/techbloghub
+DATABASE_USERNAME=admin
+DATABASE_PASSWORD=password
+
+# Admin Configuration
+ADMIN_USERNAME=admin
+ADMIN_PASSWORD=admin123
+
+# Crawler Configuration
+CRAWLER_ENABLED=true
+CRAWLER_SCHEDULE_ALL_BLOGS=0 0 3 * * ?
+CRAWLER_SCHEDULE_NEEDED_BLOGS=0 0 * * * ?
+CRAWLER_SCHEDULE_STATS=0 0 23 * * ?
+CRAWLER_STATS_ENABLED=false
+CRAWLER_TIMEOUT_SECONDS=30
+CRAWLER_USER_AGENT=TechBlogHub-Crawler/1.0
+CRAWLER_MAX_POSTS_PER_FEED=50
+
+# OpenAI Configuration (Required for LLM tagging)
+OPENAI_API_KEY=your-openai-api-key-here
+
+# Server Configuration
+SERVER_PORT=8080
+
+# Elasticsearch (Optional)
+ELASTICSEARCH_HOST=http://localhost:9200
