+++ backend/build.gradle
@@ -10,6 +10,8 @@ allprojects {
     
     repositories {
         mavenCentral()
+        maven { url 'https://repo.spring.io/milestone' }
+        maven { url 'https://repo.spring.io/snapshot' }
     }
 }
