+++ backend/output-web/build.gradle
@@ -1,12 +1,5 @@
-dependencies {
-    implementation project(':domain')
-
-    implementation 'org.springframework.boot:spring-boot-starter-webflux'
-    
-    // RSS Feed Parser
-    implementation 'com.rometools:rome:2.1.0'
-    implementation 'com.rometools:rome-utils:2.1.0'
-
-    // Test
-    testImplementation 'org.testcontainers:junit-jupiter'
+subprojects {
+    dependencies {
+        implementation project(':domain')
+    }
 }
