+++ backend/output-web/rss/src/main/resources/application-rss.yml
@@ -1,0 +1,1 @@
+# RSS Module - Environment-specific settings are managed by RssEnvironmentConfig
