+++ backend/output-web/rss/src/main/resources/rss-docker.properties
@@ -1,0 +1,4 @@
+# RSS Module - Docker Environment Configuration
+rss.http.timeout-seconds=${RSS_TIMEOUT_SECONDS:30}
+rss.http.user-agent=${RSS_USER_AGENT:TechBlogHub-Crawler/1.0}
+rss.http.max-posts-per-feed=${RSS_MAX_POSTS_PER_FEED:50}
