+++ backend/output-web/rss/src/main/resources/rss-local.properties
@@ -1,0 +1,4 @@
+# RSS Module - Local Environment Configuration
+rss.http.timeout-seconds=30
+rss.http.user-agent=TechBlogHub-Crawler/1.0
+rss.http.max-posts-per-feed=50
