+++ backend/output-web/rss/src/main/resources/rss-production.properties
@@ -1,0 +1,4 @@
+# RSS Module - Production Environment Configuration
+rss.http.timeout-seconds=${RSS_TIMEOUT_SECONDS}
+rss.http.user-agent=${RSS_USER_AGENT}
+rss.http.max-posts-per-feed=${RSS_MAX_POSTS_PER_FEED}
