+++ backend/settings.gradle
@@ -6,4 +6,6 @@ include 'admin-web'
 include 'output-persistence'
 include 'bootstrap'
 include 'output-web'
+include 'output-web:gpt'
+include 'output-web:rss'
 include 'batch'
